Moved header into components
